# openapi.json
{
  "components": {
    "schemas": {
...
      "CompareOutcomesRequest": {
        "properties": {
          "candidate_id_a": {
            "title": "Candidate Id A",
            "type": "string"
          },
          "candidate_id_b": {
            "title": "Candidate Id B",
            "type": "string"
          },
          "context": {
            "$ref": "#/components/schemas/CustomerContext"
          },
          "use_ai_models": {
            "default": true,
            "title": "Use Ai Models",
            "type": "boolean"
          }
        },
        "required": [
          "context",
          "candidate_id_a",
          "candidate_id_b"
        ],
        "title": "CompareOutcomesRequest",
        "type": "object"
      },
      "CompareOutcomesResponse": {
        "properties": {
          "candidate_a": {
            "$ref": "#/components/schemas/RankedRecommendation"
          },
          "candidate_b": {
            "$ref": "#/components/schemas/RankedRecommendation"
          },
          "request_summary": {
            "title": "Request Summary",
            "type": "object"
          },
          "winner_candidate_id": {
            "title": "Winner Candidate Id",
            "type": "string"
          }
        },
        "required": [
          "request_summary",
          "candidate_a",
          "candidate_b",
          "winner_candidate_id"
        ],
        "title": "CompareOutcomesResponse",
        "type": "object"
      },
...
      "FeedbackEvent": {
        "properties": {
          "anonymous_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anonymous Id"
          },
          "channel": {
            "$ref": "#/components/schemas/Channel",
            "default": "web"
          },
          "context_snapshot": {
            "title": "Context Snapshot",
            "type": "object"
          },
          "converted": {
            "default": false,
            "title": "Converted",
            "type": "boolean"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "event_type": {
            "default": "click",
            "title": "Event Type",
            "type": "string"
          },
          "recommendation_id": {
            "title": "Recommendation Id",
            "type": "string"
          },
          "revenue": {
            "default": 0.0,
            "title": "Revenue",
            "type": "number"
          }
        },
        "required": [
          "recommendation_id"
        ],
        "title": "FeedbackEvent",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RecommendationRequest": {
        "properties": {
          "candidates": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RecommendationCandidate"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidates"
          },
          "context": {
            "$ref": "#/components/schemas/CustomerContext"
          },
          "limit": {
            "default": 3,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "use_ai_models": {
            "default": true,
            "title": "Use Ai Models",
            "type": "boolean"
          },
          "use_llm": {
            "default": false,
            "title": "Use Llm",
            "type": "boolean"
          },
          "use_llm_explanation": {
            "default": false,
            "title": "Use Llm Explanation",
            "type": "boolean"
          }
        },
        "required": [
          "context"
        ],
        "title": "RecommendationRequest",
        "type": "object"
      },
      "RecommendationResponse": {
        "properties": {
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/RankedRecommendation"
            },
            "title": "Recommendations",
            "type": "array"
          },
          "request_summary": {
            "title": "Request Summary",
            "type": "object"
          }
        },
        "required": [
          "request_summary",
          "recommendations"
        ],
        "title": "RecommendationResponse",
        "type": "object"
      },
...
      "ScenarioRecommendationRequest": {
        "properties": {
          "limit": {
            "default": 3,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "scenario_text": {
            "maxLength": 4000,
            "minLength": 10,
            "title": "Scenario Text",
            "type": "string"
          },
          "use_ai_models": {
            "default": true,
            "title": "Use Ai Models",
            "type": "boolean"
          },
          "use_llm": {
            "default": false,
            "title": "Use Llm",
            "type": "boolean"
          },
          "use_llm_explanation": {
            "default": false,
            "title": "Use Llm Explanation",
            "type": "boolean"
          }
        },
        "required": [
          "scenario_text"
        ],
        "title": "ScenarioRecommendationRequest",
        "type": "object"
      },
      "SimulationRequest": {
        "properties": {
          "candidates": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RecommendationCandidate"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidates"
          },
          "context": {
            "$ref": "#/components/schemas/CustomerContext"
          }
        },
        "required": [
          "context"
        ],
        "title": "SimulationRequest",
        "type": "object"
      },
      "SyntheticTrainingRequest": {
        "properties": {
          "count": {
            "default": 20,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Count",
            "type": "integer"
          },
          "domain": {
            "default": "travel",
            "title": "Domain",
            "type": "string"
          }
        },
        "title": "SyntheticTrainingRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Detailed AI models for EDTA: intent, journey, TAPL, channel, outcome simulation, and ranking.",
    "title": "EDTA Full AI Models",
    "version": "2.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/compare-outcomes": {
      "post": {
        "operationId": "compare_outcomes_legacy_compare_outcomes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareOutcomesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareOutcomesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Compare Outcomes Legacy"
      }
    },
    "/context-graph": {
      "get": {
        "operationId": "get_context_graph_legacy_context_graph_get",
        "parameters": [
          {
            "in": "query",
            "name": "customer_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer Id"
            }
          },
          {
            "in": "query",
            "name": "anonymous_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Anonymous Id"
            }
          },
          {
            "in": "query",
            "name": "include_live",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Live",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Context Graph Legacy"
      }
    },
...
    "/experience-memory": {
      "get": {
        "operationId": "get_experience_memory_legacy_experience_memory_get",
        "parameters": [
          {
            "in": "query",
            "name": "customer_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer Id"
            }
          },
          {
            "in": "query",
            "name": "anonymous_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Anonymous Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Experience Memory Legacy"
      }
    },
    "/feedback": {
      "post": {
        "operationId": "feedback_legacy_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Feedback Legacy"
      }
    },
...
    "/recommend": {
      "post": {
        "operationId": "recommend_legacy_recommend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Recommend Legacy"
      }
    },
    "/recommend-from-scenario": {
      "post": {
        "operationId": "recommend_from_scenario_legacy_recommend_from_scenario_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioRecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Recommend From Scenario Legacy"
      }
    },
...
    "/simulate": {
      "post": {
        "operationId": "simulate_legacy_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Simulate Legacy"
      }
    },
    "/synthetic-training-data": {
      "post": {
        "operationId": "synthetic_training_data_legacy_synthetic_training_data_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyntheticTrainingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Synthetic Training Data Legacy"
      }
    },
    "/v1/compare-outcomes": {
      "post": {
        "operationId": "compare_outcomes_v1_v1_compare_outcomes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareOutcomesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareOutcomesResponseV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Compare Outcomes V1",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/context-graph": {
      "get": {
        "operationId": "context_graph_v1_v1_context_graph_get",
        "parameters": [
          {
            "in": "query",
            "name": "customer_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer Id"
            }
          },
          {
            "in": "query",
            "name": "anonymous_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Anonymous Id"
            }
          },
          {
            "in": "query",
            "name": "include_live",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Live",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Context Graph V1",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/v1/experience-memory": {
      "get": {
        "operationId": "experience_memory_v1_v1_experience_memory_get",
        "parameters": [
          {
            "in": "query",
            "name": "customer_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer Id"
            }
          },
          {
            "in": "query",
            "name": "anonymous_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Anonymous Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Experience Memory V1",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/feedback": {
      "post": {
        "operationId": "feedback_v1_v1_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Feedback V1",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/v1/recommend": {
      "post": {
        "operationId": "recommend_v1_v1_recommend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponseV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Recommend V1",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/recommend-from-scenario": {
      "post": {
        "operationId": "recommend_from_scenario_v1_v1_recommend_from_scenario_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioRecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponseV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Recommend From Scenario V1",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/v1/simulate": {
      "post": {
        "operationId": "simulate_v1_v1_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponseV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Simulate V1",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/synthetic-training-data": {
      "post": {
        "operationId": "synthetic_training_data_v1_v1_synthetic_training_data_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyntheticTrainingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Synthetic Training Data V1",
        "tags": [
          "v1"
        ]
      }
    },
...
  }
}